+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [NEXT] - YYYY-MM-DD
+- Modularized bounding boxes (should have no visible change)
+- Made item insertion and inventory updating server-side only
+- Added collisions (not sure if going to keep this)
+
 ## [0.0.2] - 2019-10-06
 - Initial release of Plonk for Minecraft 1.7.10
 - Place up to 4 items down